% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scEFSC.R
\name{scEFSC}
\alias{scEFSC}
\title{scEFSC}
\usage{
scEFSC(
   data, 
   n, 
   normalize = F, 
   dim1 = 5000,
   dim2 = 2000)
}
\arguments{
\item{data}{Gene expression matrix, with rows represent genes and columns represent samples.}

\item{n}{Number of clusters.}

\item{normalize}{Boolean variable indicating whether to perform scEFSC normalization and gene filtering before performing clustering.}

\item{dim1}{Number of features selected for the initial feature selection using non-negative autoencoder.}

\item{dim2}{Number of features for feature selections using the unsupervised feature selection algorithm.}
}
\value{
List of labels.
}
\description{
The main function to perform clustering.
}
